+++ app/src/main/java/com/example/movie/model/Movie.kt
@@ -1,6 +1,5 @@
-package com.example.movie
+package com.example.movie.model
 
-import android.os.Parcel
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
 import kotlinx.parcelize.Parcelize
