+++ app/src/main/java/com/example/movie/model/RetrofitClient.kt
@@ -1,8 +1,7 @@
-package com.example.movie
+package com.example.movie.model
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.create
 
 object RetrofitClient {
