+++ app/src/main/java/com/example/movie/view/DetailFragment.kt
@@ -1,4 +1,4 @@
-package com.example.movie
+package com.example.movie.view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,9 +6,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.navArgs
+import com.example.movie.model.Movie
 import com.example.movie.databinding.FragmentDetailBinding
 import com.squareup.picasso.Picasso
-import kotlin.Result
 
 
 class DetailFragment : Fragment() {
