+++ app/src/main/java/com/example/movie/view/MainFragment.kt
@@ -1,23 +1,18 @@
-package com.example.movie
+package com.example.movie.view
 
-import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
+import com.example.movie.model.Movie
+import com.example.movie.model.MovieList
 import com.example.movie.databinding.FragmentMainBinding
+import com.example.movie.model.Common
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-import kotlin.Result
 import kotlin.coroutines.CoroutineContext
 
 class MainFragment : Fragment(), CoroutineScope {
@@ -47,10 +42,10 @@ class MainFragment : Fragment(), CoroutineScope {
         launch {
             movie = Common.getPostApi().getMoviesList()
 
-            adapter=MyMovieAdapter(list=movie.results)
+            adapter= MyMovieAdapter(list=movie.results)
             binding.recyclerView.adapter =adapter
 
-            adapter.movieClick=object:MyMovieAdapter.MovieItemClick{
+            adapter.movieClick=object: MyMovieAdapter.MovieItemClick {
                 override fun movieItemClick(item: Movie) {
                     val action = MainFragmentDirections.actionMainFragment2ToDetailFragment2(item)
                     findNavController().navigate(action)
