+++ app/src/main/res/layout/second_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SecondFragment">
+    tools:context=".view.SecondFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
